fix/#88: saveProfileInfor->profile 파일명 변경 및 header 제거
